Delete some old code, and remove single function/class files.
