+++ readthedocs/search/apps.py
@@ -7,4 +7,4 @@ class SearchConfig(AppConfig):
     name = 'readthedocs.search'
 
     def ready(self):
-        from .signals import index_html_file, remove_html_file
+        from .signals import index_html_file, remove_html_file, index_project, remove_project  # noqa
